+++ dojo/engagement/__init__.py
@@ -1,0 +1,1 @@
+import dojo.engagement.admin  # noqa: F401
